+++ back_end/demo/src/main/java/com/avelina_anton/bzhch/smart_house/demo/repositories/UsersRepository.java
@@ -1,3 +1,4 @@
+
 package com.avelina_anton.bzhch.smart_house.demo.repositories;
 
 import com.avelina_anton.bzhch.smart_house.demo.models.User;
@@ -5,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository<User, Integer> {
+public interface UsersRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
     Optional<User> findByName(String name);
     boolean existsByName(String name);
